MARK_I/backend_python/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
frozenlist==1.5.0
h11==0.14.0
httptools==0.6.4
idna==3.10
livekit-api==1.0.2
livekit-protocol==1.0.2
multidict==6.4.3
propcache==0.3.1
protobuf==6.30.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
PyNaCl==1.5.0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
telnyx==2.1.5
types-protobuf==5.29.1.20250403
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
